Minor fixes throughout code.  The first issue was that a number of locations in the code used uninitialized memory values, which worked on some systems thanks to GCC's approach of initializing everything to zero.  Second, I cleaned up numerous locations in the code where we set variables but never used them; while these weren't bugs these changes cut down dramatically on compiler warnings.

+++ examples/cvode/fcmix_serial/fcvAdvDiff_bnd.f
@@ -170,7 +170,7 @@ C
 C Compute max-norm of array U
       IMPLICIT NONE
 C
-      INTEGER I, N
+      INTEGER*8 I, N
       DOUBLE PRECISION U(*), UNORM, TEMP
 C
       TEMP = 0.0D0
